User Model Created

+++ app/Http/Controllers/InteractionController.php
@@ -7,14 +7,14 @@ use App\Http\Requests\StoreInteractionRequest;
 use App\Http\Requests\UpdateInteractionRequest;
 use App\Models\Customer;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class InteractionController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    use AuthorizesRequests;
     public function index()
     {
+        $this->authorize('viewAny',Interaction::class);
         $columns = ['customer_id','type','subject','details','interaction_date'];
         $query = Interaction::with('user','customer')->latest();
         if (request()->filled('search')) {
@@ -45,6 +45,7 @@ class InteractionController extends Controller
      */
     public function create()
     {
+        $this->authorize('create',Interaction::class);
         $users = User::orderBy('first_name', 'asc')->get();
         $customers = Customer::orderBy('first_name', 'asc')->get();
         return view('interactions.create', compact('users','customers'));
@@ -55,6 +56,7 @@ class InteractionController extends Controller
      */
     public function store(StoreInteractionRequest $request)
     {
+        $this->authorize('create',Interaction::class);
         Interaction::create($request->validated());
 
         return redirect()->route('interactions.index')->with('success','Interaction created successfully');
@@ -65,6 +67,7 @@ class InteractionController extends Controller
      */
     public function show(Interaction $interaction)
     {
+        $this->authorize('view',$interaction);
         return view('interactions.show',compact('interaction'));
     }
 
@@ -73,6 +76,7 @@ class InteractionController extends Controller
      */
     public function edit(Interaction $interaction)
     { 
+        $this->authorize('update',$interaction);
         $users = User::orderBy('first_name', 'asc')->get();
         $customers = Customer::orderBy('first_name', 'asc')->get();
         return view('interactions.edit',compact('interaction','users','customers'));
@@ -83,6 +87,7 @@ class InteractionController extends Controller
      */
     public function update(UpdateInteractionRequest $request, Interaction $interaction)
     {
+        $this->authorize('update',$interaction);
         $interaction->update($request->validated());
 
         return redirect()->route('interactions.index')->with('success','Interaction edited successfully');
@@ -93,19 +98,22 @@ class InteractionController extends Controller
      */
     public function destroy(Interaction $interaction)
     {
+        $this->authorize('delete',$interaction);
         $interaction->delete();
 
         return redirect()->route('interactions.index')->with('success','Interaction deleted successfully');
     }
-    public function restore($id)
+    public function restore(Interaction $interaction,$id)
 {
+    $this->authorize('restore',$interaction);
     $interaction = Interaction::onlyTrashed()->findOrFail($id);
     $interaction->restore();
 
     return redirect()->route('interactions.index')->with('success', 'Interaction restored successfully');
 }
-public function forceDelete($id)
+public function forceDelete(Interaction $interaction,$id)
 {
+    $this->authorize('forceDelete',$interaction);
     $interaction = Interaction::onlyTrashed()->findOrFail($id);
     $interaction->forceDelete();
